network, vmispec: Move GetNamespaceAndNetworkName to vmispec package
This way we can use it in a more generic way,
i.e on packages which the network package itself imports,
without creating cycling dependency.
