+++ pkg/network/vmispec/BUILD.bazel
@@ -10,7 +10,10 @@ go_library(
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/vmispec",
     visibility = ["//visibility:public"],
-    deps = ["//staging/src/kubevirt.io/api/core/v1:go_default_library"],
+    deps = [
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/precond:go_default_library",
+    ],
 )
 
 go_test(
@@ -29,5 +32,6 @@ go_test(
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
     ],
 )
